+++ vitest.node.config.ts
@@ -85,10 +85,14 @@ export default defineConfig({
     },
     experimental: {
       openTelemetry: {
-        enabled: process.env.VITEST_OTEL_ENABLED === 'true',
+        enabled:
+          process.env.VITEST_OTEL_ENABLED === 'true' &&
+          // not set in dependabot PRs
+          !!process.env.OTEL_EXPORTER_OTLP_ENDPOINT,
         sdkPath: './vitest.node.otel.js',
       },
     },
+    retry: process.env.CI ? 2 : 0,
   },
   resolve: {
     alias: {